pyproject.toml
[project]
name = "agentauth"
version = "0.2.1"
description = "Automated authentication for web agents"
readme = "README.md"
authors = [
    { name = "Cole Hecht" }
]
requires-python = ">=3.11"
dependencies = [
    "browser-use>=0.1.36",
    "browserbase>=1.1.0",
    "imap-tools>=1.10.0",
    "onepassword-sdk>=0.1.7",
    "pyotp>=2.9.0",
    "python-dotenv>=1.0.1",
    "langchain>=0.1.0",
    "structlog>=25.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
